% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup_funs.R
\name{as_matrix}
\alias{as_matrix}
\title{Rcpp version's as.matrix}
\usage{
as_matrix(mat)
}
\arguments{
\item{mat}{Sparse matrix.}
}
\value{
Matrix.
}
\description{
Transform sparse matrix to matrix.
}
\examples{
aa<-matrix(seq(1,6),nrow=2,ncol=3)
qq<-as(as.matrix(aa), "dgCMatrix")
all.equal(unname(as_matrix(qq)),unname(as.matrix(qq)))
}
