% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asMatrix}
\alias{asMatrix}
\title{Rcpp version: as.matrix}
\usage{
asMatrix(rp, cp, z, nrows, ncols)
}
\arguments{
\item{rp}{vector}

\item{cp}{vector}

\item{z}{vector}

\item{nrows}{nrows}

\item{ncols}{ncols}
}
\value{
Matrix object in R.
}
\description{
Rcpp version: as.matrix
}
\details{
Rcpp version: as.matrix
}
\examples{
data("EXAMPLE_DATA_list")
#Should not run by the users, it is used in prior estimation.
\dontrun{
}
}
