% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayNorm.r
\name{bayNorm}
\alias{bayNorm}
\title{A wrapper function of prior estimation and bayNorm function}
\usage{
bayNorm(Data, BETA_vec, S = 20, Para = T, NCores = 5, FIX_MU = T,
  GR = F, Conditions = NULL, BB_SIZE = T)
}
\arguments{
\item{Data:}{A matrix of single-cell expression where rows are genes and columns are samples (cells). This object should be of class matrix rather than data.frame.}

\item{BETA_vec:}{A vector of capture efficiencies of cells.}

\item{S:}{The number of samples you would like to genrate. Default is 20.}

\item{Para:}{If TRUE, 5 cores will be used for parallelization.}

\item{NCores:}{number of cores to use, default is 5.}

\item{FIX_MU:}{If TRUE, then 1D optimization, otherwise 2D optimization (slow).}

\item{GR:}{If TRUE, the gradient function will be used in optimization. However since the gradient function itself is very complicated, it does not help too much in speeding up. Default is FALSE.}

\item{Conditions:}{Default is NULL.}

\item{BB_SIZE:If}{TRUE, estimate BB size, and then use it for adjusting MME SIZE. Use the adjusted MME size for bayNorm. Defaut is TRUE.}
}
\value{
A list of objects.
}
\description{
Input raw data and a vector of capture efficiencies of cells. You may
also need to specify the condition of cells.
}
