% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sup_funs.R
\name{Check_input}
\alias{Check_input}
\title{Check input}
\usage{
Check_input(Data)
}
\arguments{
\item{Data}{Input data.}
}
\value{
Matrix (of type matrix in R).
}
\description{
Check input
}
\examples{
aa<-matrix(seq(1,6),nrow=2,ncol=3)
Check_input(aa)

}
