% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{DownSampling}
\alias{DownSampling}
\title{Binomial downsampling}
\usage{
DownSampling(Data, BETA_vec)
}
\arguments{
\item{Data:}{raw count Data}

\item{BETA_vec:}{A vector of capture efficiencies of cells}
}
\value{
A matrix of binomial downsampling data.
}
\description{
For each element in the \code{Data}, randomly generate a number using Binomial distribution with probability equal to the specific capture efficiency.
}
