% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GradientFun_1D}
\alias{GradientFun_1D}
\title{GradientFun_1D}
\usage{
GradientFun_1D(SIZE, MU, m_observed, BETA)
}
\arguments{
\item{SIZE:}{SIZE}

\item{MU:}{MU}

\item{m_observed:}{m_observed}

\item{BETA:}{BETA}
}
\value{
GradientFun_1D
}
\description{
GradientFun_1D
}
