% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GradientFun_2D}
\alias{GradientFun_2D}
\title{GradientFun_2D}
\usage{
GradientFun_2D(SIZE_MU, m_observed, BETA)
}
\arguments{
\item{m_observed:}{m_observed}

\item{BETA:}{BETA}
}
\value{
GradientFun_2D
}
\description{
GradientFun_2D
}
