% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GradientFun_NB_1D}
\alias{GradientFun_NB_1D}
\title{GradientFun_NB_1D}
\usage{
GradientFun_NB_1D(SIZE, MU, m_observed, BETA)
}
\arguments{
\item{SIZE}{size}

\item{MU}{mu}

\item{m_observed}{one observed count}

\item{BETA}{The corresponding capture efficiency}
}
\value{
GradientFun_NB_1D
}
\description{
First derivative of marginal distribution
with respect to size.
}
\examples{
data("EXAMPLE_DATA_list")
#Should not run by the users, it is used in prior estimation.
\dontrun{
}
}
