% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GradientFun_NB_2D}
\alias{GradientFun_NB_2D}
\title{GradientFun_NB_2D}
\usage{
GradientFun_NB_2D(SIZE_MU, m_observed, BETA)
}
\arguments{
\item{m_observed}{one observed count}

\item{BETA}{The corresponding capture efficiency}

\item{SIZE}{size}

\item{MU}{mu}
}
\value{
GradientFun_NB_2D
}
\description{
First derivative of marginal distribution
with respect to size.
}
\examples{
data("EXAMPLE_DATA_list")
#Should not run by the users.
\dontrun{
}
}
