% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Main_Bay}
\alias{Main_Bay}
\title{Main_Bay}
\usage{
Main_Bay(Data, BETA_vec, size, mu, S, thres, Mean_depth)
}
\arguments{
\item{Data:}{raw count Data}

\item{BETA_vec:}{A vector of capture efficiencies of cells}

\item{size:}{A vector of size}

\item{mu:}{A vector of mu}

\item{S:}{number of samples that you want to generate}
}
\value{
bayNorm
}
\description{
bayNorm
there are leading NA, marking the leadings NA as TRUE and
everything else as FALSE.
}
