% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Main_NB_Bay}
\alias{Main_NB_Bay}
\title{Main_NB_Bay}
\usage{
Main_NB_Bay(Data, BETA_vec, size, mu, S, thres)
}
\arguments{
\item{Data}{raw count Data}

\item{BETA_vec}{A vector of capture efficiencies of cells}

\item{size}{A vector of size}

\item{mu}{A vector of mu}

\item{S}{Draw S samples from posterior
distribution to form 3D array}

\item{thres}{useless parameter}
}
\value{
bayNorm normalized data
}
\description{
This function is used to draw S samples from
Negative Binomial distribution for each gene in each cell.
}
\examples{
data("EXAMPLE_DATA_list")
\dontrun{
Norm_3D_array<-Main_NB_Bay(Data=EXAMPLE_DATA_list$inputdata,
BETA_vec = EXAMPLE_DATA_list$inputbeta,
size=EXAMPLE_DATA_list$size,mu=EXAMPLE_DATA_list$mu,
S=20,thres=10000000)
}
}
