% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MarginalF_1D}
\alias{MarginalF_1D}
\title{MarginalF_1D}
\usage{
MarginalF_1D(SIZE, MU, m_observed, BETA)
}
\arguments{
\item{SIZE}{size}

\item{MU}{mu}

\item{m_observed}{one observed count}

\item{BETA}{The corresponding capture efficiency}
}
\value{
Marginal likelihood
}
\description{
MarginalF_1D
}
\examples{
data("EXAMPLE_DATA_list")
#Should not run by the users.
\dontrun{
}
}
