% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MarginalF_NB_2D}
\alias{MarginalF_NB_2D}
\title{MarginalF_NB_2D}
\usage{
MarginalF_NB_2D(SIZE_MU, m_observed, BETA)
}
\arguments{
\item{SIZE_MU}{a vector of two elements (size,mu)}

\item{m_observed}{m_observed}

\item{BETA}{Corresponding capture efficiency}
}
\value{
Marginal likelihood
}
\description{
Mariginal distribution as a function
of both size and mu.
}
\examples{
data("EXAMPLE_DATA_list")
#Should not run by the users.
\dontrun{
}
}
