% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisygene_detection.r
\name{NOISY_FUN}
\alias{NOISY_FUN}
\title{Noisy gene detection}
\usage{
NOISY_FUN(bay_array_N, bay_array_C, plot.out = F)
}
\arguments{
\item{bay_array_N}{A 2D matrix or 3D array of normalized data(real cells).}

\item{bay_array_C}{A 2D matrix or 3D array of normalized data(synthetic control).}
}
\description{
Input raw data and a vector of capture efficiencies of cells. You can
also need to specify the condition of cells.
}
\details{
Noisy gene detection
}
