% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisygene_detection.r
\name{noisy_gene_detection}
\alias{noisy_gene_detection}
\title{A wrapper function of synthetic control generation, bayNorm on both real cell data and synthetic controls and noisy gene detection.}
\usage{
noisy_gene_detection(Data, BETA_vec, mode_version = F, S = 20,
  parallel = T, NCores = 5, FIX_MU = T, GR = F, BB_SIZE = T,
  verbose = T, plot.out = F)
}
\arguments{
\item{Data}{A matrix of single-cell expression where rows are genes and columns are samples (cells). This object should be of class matrix rather than data.frame.}

\item{BETA_vec}{A vector of capture efficiencies of cells.}

\item{mode_version}{If TRUE, bayNorm return mode version normalized data which is of 2D matrix instead of 3D array. Default is FALSE.}

\item{S}{The number of samples you would like to generate from estimated posterior distribution (The third dimension of 3D array). Default is 20. S needs to be specified if \code{mode_version}=FALSE.}

\item{parallel}{If TRUE, 5 cores will be used for parallelization.}

\item{NCores}{number of cores to use, default is 5. This will be used to set up a parallel environment using either MulticoreParam (Linux, Mac) or SnowParam (Windows) with NCores using the package BiocParallel.}

\item{FIX_MU}{Whether fix mu when estimating parameters by maximizing marginal distribution. If TRUE, then 1D optimization, otherwise 2D optimization (slow).}

\item{GR}{If TRUE, the gradient function will be used in optimization. However since the gradient function itself is very complicated, it does not help too much in speeding up. Default is FALSE.}

\item{BB_SIZE}{If TRUE, estimate BB size, and then use it for adjusting MME SIZE. Use the adjusted MME size for bayNorm. Default is TRUE.}

\item{verbose}{print out status messages. Default is TRUE.}

\item{Conditions}{No need to specify.}

\item{UMI_sffl}{No need to specify. Currently this function is for UMI based data.}

\item{Prior_type}{No need to specify. Default is NULL. If \code{Conditions} is NULL, priors are estimated based on all cells.}
}
\value{
List of objects. The first element in the list is the adjusted P-values (noisy genes detection).
}
\description{
A wrapper function of synthetic control generation, bayNorm on both real cell data and synthetic controls and noisy gene detection.
}
\details{
A wrapper function of synthetic control generation, bayNorm on both real cell data and synthetic controls and noisy gene detection.
}
