% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIOR_FUNCTIONS.R
\name{Prior_fun}
\alias{Prior_fun}
\title{A wrapper function of EstPrior and AdjustSIZE_fun}
\usage{
Prior_fun(Data, BETA_vec, parallel = T, NCores = 5, FIX_MU = T, GR = F,
  BB_SIZE = T)
}
\arguments{
\item{Data:}{A matrix of single-cell expression where rows are genes and columns are samples (cells). This object should be of class matrix rather than data.frame.}

\item{BETA_vec:}{A vector of capture efficiencies of cells.}

\item{parallel:}{If TRUE, 5 cores will be used for parallelization. Defaut is TRUE.}

\item{NCores:}{number of cores to use, default is 5. This will be used to set up a parallel environment using either MulticoreParam (Linux, Mac) or SnowParam (Windows) with NCores using the package BiocParallel.}

\item{FIX_MU:}{If TRUE, then 1D optimization, otherwise 2D optimization (slow). Defaut is TRUE.}

\item{GR:}{If TRUE, the gradient function will be used in optimization. However since the gradient function itself is very complicated, it does not help too much in speeding up. Default is FALSE.}

\item{BB_SIZE:If}{TRUE, estimate BB size, and then use it for adjusting MME SIZE. Use the adjusted MME size for bayNorm. Defaut is TRUE.}
}
\value{
A list of objects.
}
\description{
Input raw data and a vector of capture efficiencies of cells.
}
