% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIOR_FUNCTIONS.r
\name{AdjustSIZE_fun}
\alias{AdjustSIZE_fun}
\title{Adjust MME size}
\usage{
AdjustSIZE_fun(BB_SIZE, MME_MU, MME_SIZE)
}
\arguments{
\item{BB_SIZE}{size estimated from \code{BB_Fun}.}

\item{MME_MU}{mu estimated from EstPrior.}

\item{MME_SIZE}{size estimated from EstPrior.}
}
\value{
MME_SIZE_adjust: A vector of estimated size.
Adjusted MME_SIZE based on BB_SIZE
(size estimated by maximizing marginal distribution)
}
\description{
Adjust MME estimated size according to size
estimated through maximizing marginal distirbution
(\code{BB_SIZE}). By using maximizing marginal
distribution,
a log of estimated size cannot converge quickly.
However according to our simulation,
the trend of \code{BB_SIZE}
is very close to the true size. Hence it is better to use
adjusted MME size in bayNorm.
}
\examples{
data('EXAMPLE_DATA_list')
\dontrun{
}
}
