% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIOR_FUNCTIONS.R
\name{AdjustSIZE_fun}
\alias{AdjustSIZE_fun}
\title{Adjust MME size}
\usage{
AdjustSIZE_fun(BB_SIZE, MME_MU, MME_SIZE)
}
\arguments{
\item{BB_SIZE}{size estimated from BB_Fun or BB_Fun_1D}

\item{MME_MU}{mu estimated from EstPrior.}

\item{MME_SIZE}{size estimated from EstPrior.}
}
\value{
MME_SIZE_adjust A vector of estimated size. Adjusted MME_SIZE based on BB_size
}
\description{
Adjust MME size
}
