% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIOR_FUNCTIONS.r
\name{EstPrior}
\alias{EstPrior}
\title{Estimate size and mu for Negative Binomial distribution
for each gene using MME method}
\usage{
EstPrior(Data, parallel = FALSE, NCores = 5, verbose = TRUE)
}
\arguments{
\item{Data}{A matrix of single-cell expression where rows
are genes and columns are samples (cells). \code{Data}
can be of class \code{SummarizedExperiment} (the
assays slot contains the expression matrix and
is named "Counts") or just matrix.}

\item{parallel}{If TRUE, \code{NCores} cores will be
used for parallelization. Default is TRUE.}

\item{NCores}{number of cores to use, default is 5.
This will be used to set up a parallel
environment using either MulticoreParam (Linux, Mac)
or SnowParam (Windows) with NCores
using the package BiocParallel.}

\item{verbose}{print out status messages. Default is TRUE.}
}
\value{
List containing estimated mu and
size for each gene.
}
\description{
Input raw data and return
estimated size and mu for each gene using the MME method.
}
\details{
mu and size are two parameters of the prior that
need to be specified for each gene in bayNorm.
They are parameters of negative binomial distribution.
The variance is \eqn{mu + mu^2/size} in this parametrization.
}
\examples{
data('EXAMPLE_DATA_list')
#Return estimated mu and size for each gene using MME method.
MME_est<-EstPrior(Data=EXAMPLE_DATA_list$inputdata[,seq(1,30)],
verbose=TRUE)
}
