% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRIOR_FUNCTIONS.R
\name{EstPrior}
\alias{EstPrior}
\title{Estimate size and mu for NB distribution for each gene using MME method}
\usage{
EstPrior(Data)
}
\arguments{
\item{Data:}{A matrix of single-cell expression where rows are genes and columns are samples (cells). This object should be of class matrix rather than data.frame.}
}
\value{
Estimated size and mu for each gene.
}
\description{
Input raw data and return estimated size and mu for each gene.
}
