% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXAMPLE_DATA_list.r
\docType{data}
\name{EXAMPLE_DATA_list}
\alias{EXAMPLE_DATA_list}
\title{A subset of Grun et al (2014) data: 2i samples}
\format{A list: EXAMPLE_DATA_list[[1]]: inputdata, EXAMPLE_DATA_list[[2]]: inputbeta (a vector of probabilities with length equal  to the number of cells), EXAMPLE_DATA_list[[3]]: mu (MME method estimated mean expression for each gene), EXAMPLE_DATA_list[[4]]: size (adjusted MME size for each gene).}
\usage{
EXAMPLE_DATA_list
}
\description{
Small extract (20 genes and 74 cells) from the
}
\examples{
data(EXAMPLE_DATA_list)
}
\references{
Grun, Kester and van Oudenaarden (2014). Nature Methods.
}
\keyword{datasets}
