% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXAMPLE_DATA_list.r
\docType{data}
\name{EXAMPLE_DATA_list}
\alias{EXAMPLE_DATA_list}
\title{Example datasets for bayNorm}
\format{list}
\usage{
EXAMPLE_DATA_list
}
\description{
A list contains small example data.
inputdata: A UMI based scRNAseq data (1011 genes and 96 cells).
The corresponding capture efficiencies (inputbeta), prior
estimates mu (mu) and size (size) are also provided.
}
\examples{
data(EXAMPLE_DATA_list)
}
\keyword{datasets}
