% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GradientFun_2D}
\alias{GradientFun_2D}
\title{GradientFun_2D}
\usage{
GradientFun_2D(SIZE_MU, m_observed, BETA)
}
\arguments{
\item{SIZE_MU}{a vector of two elements (size,mu)}

\item{m_observed}{one observed count}

\item{BETA}{The corresponding capture efficiency}
}
\value{
GradientFun_2D
}
\description{
GradientFun_2D
}
\examples{
data("EXAMPLE_DATA_list")
\dontrun{
}


}
