% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GradientFun_NB_2D}
\alias{GradientFun_NB_2D}
\title{GradientFun_NB_2D}
\usage{
GradientFun_NB_2D(SIZE_MU, m_observed, BETA)
}
\arguments{
\item{SIZE_MU}{a vector of two elements (size,mu)}

\item{m_observed}{a vector of observed counts}

\item{BETA}{The corresponding capture efficiency}
}
\value{
GradientFun_NB_2D
}
\description{
First derivative of marginal distribution
with respect to size and mu.
}
\examples{
data("EXAMPLE_DATA_list")
#Should not run by the users.
\dontrun{
}
}
