% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Main_mean_NB_Bay}
\alias{Main_mean_NB_Bay}
\title{Mean_NB_Bay}
\usage{
Main_mean_NB_Bay(Data, BETA_vec, size, mu, S, thres)
}
\arguments{
\item{Data}{raw count Data}

\item{BETA_vec}{A vector of capture efficiencies of cells}

\item{size}{A vector of size}

\item{mu}{A vector of mu}

\item{S}{number of samples that you want to generate}

\item{thres}{useless parameter}
}
\value{
bayNorm normalized data
}
\description{
bayNorm
If the observed count is above 500,
then we use normal distribution to
approximate binomial distribution.
}
\examples{
data("EXAMPLE_DATA_list")
\dontrun{
data("EXAMPLE_DATA_list")
Norm_2D_matrix<-Main_mean_Bay(Data=EXAMPLE_DATA_list$inputdata,
BETA_vec = EXAMPLE_DATA_list$inputbeta,
size=EXAMPLE_DATA_list$size,mu=EXAMPLE_DATA_list$mu,
S=20,thres=10000000)
}
}
