% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Main_mode_Bay}
\alias{Main_mode_Bay}
\title{Mode_Bay}
\usage{
Main_mode_Bay(Table, Beta_origin, size, M_ave_ori, S, thres, Mean_depth)
}
\arguments{
\item{Table:}{raw count table}

\item{Beta_origin:}{A vector of capture efficiencies of cells}

\item{size:}{A vector of size}

\item{M_ave_ori:}{A vector of mu}

\item{S:}{number of samples that you want to generate}
}
\value{
bayNorm
}
\description{
bayNorm
there are leading NA, marking the leadings NA as TRUE and
everything else as FALSE.
}
