% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MarginalF_1D}
\alias{MarginalF_1D}
\title{MarginalF_1D}
\usage{
MarginalF_1D(SIZE, MU, m_observed, BETA)
}
\arguments{
\item{size:}{A vector of size}

\item{mu:}{A vector of mu}
}
\value{
Marginal likelihood
}
\description{
MarginalF_1D
}
