% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MarginalF_2D}
\alias{MarginalF_2D}
\title{MarginalF_2D}
\usage{
MarginalF_2D(SIZE_MU, m_observed, BETA)
}
\arguments{
\item{size:}{A vector of size}

\item{mu:}{A vector of mu}
}
\value{
Marginal likelihood
}
\description{
MarginalF_2D
}
